user test added

+++ src/modules/user/user.repository.ts
@@ -1,5 +1,7 @@
 import prisma from "../../db/connectDB";
 import { Prisma } from "../../../generated/prisma/client";
+import AppError from "../../errors/appError";
+import { status } from "../../utils/status";
 
 class UserRepository {
   async findAllUsers() {
@@ -11,14 +13,26 @@ class UserRepository {
   }
 
   async findUserById(id: string) {
-    return prisma.user.findUnique({ where: { id } });
+    const findUser = await prisma.user.findUnique({ where: { id } });
+    if (!findUser) {
+      throw new AppError("user not found", status.NOT_FOUND);
+    }
+    return findUser;
   }
 
   async updateUser(id: string, data: Prisma.UserUpdateInput) {
-    return prisma.user.update({ where: { id }, data });
+    const findUser = await prisma.user.findUnique({ where: { id } });
+    if (!findUser) {
+      throw new AppError("user not found", status.NOT_FOUND);
+    }
+
+    return prisma.user.update({ where: { id: findUser.id }, data });
   }
 
   async deleteUser(id: string) {
+    if (!id) {
+      throw new AppError("id is required", status.BAD_REQUEST);
+    }
     return prisma.user.delete({ where: { id } });
   }
 }
