+++ src/modules/user/user.services.ts
@@ -1,13 +1,12 @@
+import { Prisma } from "../../../generated/prisma/client";
 import AppError from "../../errors/appError";
 import { status } from "../../utils/status";
-import { Prisma } from "../../../generated/prisma/client";
 import UserRepository from "./user.repository";
 
 class UserService {
   constructor(private userRepo: UserRepository) {}
 
   async createUser(data: Prisma.UserCreateInput) {
-    console.log("data", data);
     return this.userRepo.createUser(data);
   }
 
@@ -17,19 +16,43 @@ class UserService {
 
   async findUserById(id: string) {
     if (!id) {
-      throw new AppError("Enter a valid user Id", status.NOT_FOUND);
+      throw new AppError("Enter a valid user id.", status.BAD_REQUEST);
     }
-    return this.userRepo.findUserById(id);
+
+    const user = await this.userRepo.findUserById(id);
+
+    if (!user) {
+      throw new AppError("User not found.", status.NOT_FOUND);
+    }
+
+    return user;
   }
 
   async updateUser(id: string, data: Prisma.UserUpdateInput) {
     if (!id) {
-      throw new AppError("Enter a valid user Id", status.NOT_FOUND);
+      throw new AppError("Enter a valid user id.", status.BAD_REQUEST);
     }
+
+    const user = await this.userRepo.findUserById(id);
+
+    if (!user) {
+      throw new AppError("User not found.", status.NOT_FOUND);
+    }
+
     return this.userRepo.updateUser(id, data);
   }
 
   async deleteUser(id: string) {
+    if (!id) {
+      throw new AppError("Enter a valid user id.", status.BAD_REQUEST);
+    }
+
+    const user = await this.userRepo.findUserById(id);
+
+    if (!user) {
+      throw new AppError("User not found.", status.NOT_FOUND);
+    }
+
     return this.userRepo.deleteUser(id);
   }
 }
